requirements.txt
PyQt6>=6.6
sqlalchemy>=2.0
passlib[bcrypt]>=1.7
pyserial>=3.5
reportlab>=4.0
openpyxl>=3.1
qrcode[pil]>=7.4
Pillow>=10.0
python-dotenv>=1.0
PyInstaller>=6.0
